+++ src/constants/api.ts
@@ -13,7 +13,7 @@ const END_POINTS = {
   CREATE_CLUB: '/clubs',
   INVITE_LINK: '/club/invite', // TODO: API 명세서 나올시, 수정 필요
   POST_EVENT_APPLY: '/events/apply',
-  CLUB_EVENTS: ({ clubId = 1 }: { clubId: number }) => `/clubs/${clubId}/events`,
+  CLUB_EVENTS: ({ clubId }: { clubId: number }) => `/clubs/${clubId}/events`,
 } as const;
 
 export { SPACECLUB_BASE_URL, NETWORK_TIMEOUT, END_POINTS };
